+++ app/src/main/java/com/timmytruong/simplelauncher/data/Event.kt
@@ -1,4 +1,4 @@
-package com.timmytruong.habit_launcher.data
+package com.timmytruong.simplelauncher.data
 
 sealed class Event {
     data class NavigateToPackage(val packageName: CharSequence): Event()
