+++ app/src/main/java/com/timmytruong/simplelauncher/ui/main/composable/AppList.kt
@@ -1,4 +1,4 @@
-package com.timmytruong.habit_launcher.ui.main.composable
+package com.timmytruong.simplelauncher.ui.main.composable
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.lazy.LazyColumn
@@ -7,7 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.timmytruong.habit_launcher.ui.main.MainActivityViewModel
+import com.timmytruong.simplelauncher.ui.main.MainActivityViewModel
 
 @Composable
 internal fun AppList(
