+++ app/src/main/java/com/timmytruong/simplelauncher/ui/menu/composable/MenuBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.timmytruong.habit_launcher.ui.menu.composable
+package com.timmytruong.simplelauncher.ui.menu.composable
 
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ModalBottomSheet
@@ -6,8 +6,8 @@ import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.timmytruong.habit_launcher.data.AppInfo
-import com.timmytruong.habit_launcher.ui.menu.MenuBottomSheetViewModel
+import com.timmytruong.simplelauncher.data.AppInfo
+import com.timmytruong.simplelauncher.ui.menu.MenuBottomSheetViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
