+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">habit-launcher</string>
+    <string name="app_name">simple-launcher</string>
     <string name="mainActivity">Main Activity</string>
     <string name="app_icon">App Icon</string>
 </resources>
