+++ buildSrc/src/main/java/Config.kt
@@ -1,5 +1,5 @@
 object Config {
-    const val APP_ID = "com.timmytruong.habit_launcher"
+    const val APP_ID = "com.timmytruong.simplelauncher"
     const val VERSION_NAME = "1.0"
     const val VERSION_CODE = 1
